Move db files into their own folder structure
