+++ src/main.py
@@ -4,8 +4,8 @@ from fastapi import (
 )
 from fastapi.middleware.cors import CORSMiddleware
 import uvicorn
-import db
-from db_helpers import row2dict
+import db.db_operations as db_operations
+from db.db_helpers import row2dict
 from pydantic import BaseModel, ValidationError, validator
 from datetime import date
 import traceback
@@ -49,9 +49,9 @@ async def root():
 
 @app.post("/get_events_by_month")
 async def get_events_by_month(req: GetEventsByMonth):
-    with db.session_scope() as session:
-        events = db.get_events_by_month(session,req.month,req.year)
-        tags = db.get_event_tags(session,events)
+    with db_operations.session_scope() as session:
+        events = db_operations.get_events_by_month(session,req.month,req.year)
+        tags = db_operations.get_event_tags(session,events)
         return {
             'events': row2dict(events),
             'tags': tags
@@ -59,9 +59,9 @@ async def get_events_by_month(req: GetEventsByMonth):
     
 @app.post("/get_events_by_date")
 async def get_events_by_date(req: GetEventsByDate):
-    with db.session_scope() as session:
-        events = db.get_events_by_date(session,req.from_date,req.include_description)
-        tags = db.get_event_tags(session,events)
+    with db_operations.session_scope() as session:
+        events = db_operations.get_events_by_date(session,req.from_date,req.include_description)
+        tags = db_operations.get_event_tags(session,events)
         return {
             'events': row2dict(events),
             'tags': tags
@@ -92,14 +92,14 @@ async def digest(req: EmailModel):
     verified = parsed.sender.email.domain.lower() == "mit.edu" # could be improved?
     sender_email = str(parsed.sender.email).lower()
     if verified:
-        with db.session_scope() as session:
-            user_id = db.add_user(session, sender_email)
+        with db_operations.session_scope() as session:
+            user_id = db_operations.add_user(session, sender_email)
             club_id = None
             location = None
             for location in parsed.locations: break
             link = None
 
-            event_id = db.add_event(
+            event_id = db_operations.add_event(
                 session,
                 parsed.thread_topic,
                 user_id,
